+++ bmsdna/lakeapi/core/partition_utils.py
@@ -1,0 +1,29 @@
+from typing import List
+import os
+
+from bmsdna.lakeapi.core.config import BasicConfig
+
+from bmsdna.lakeapi.core.types import Param
+
+
+def should_hide_colname(name: str):
+    return name.startswith("_") or "_md5_prefix_" in name or "_xxhash64_prefix_" in name or "_md5_mod_" in name
+
+
+def _with_implicit_parameters(paramslist: List[Param], file_type: str, basic_config: BasicConfig, uri: str):
+    if file_type == "delta":
+        delta_uri = os.path.join(
+            basic_config.data_path,
+            uri,
+        )
+        from deltalake import DeltaTable
+
+        part_cols = DeltaTable(delta_uri).metadata().partition_columns
+        if part_cols and len(part_cols) > 0:
+            all_names = [(p.colname or p.name).lower() for p in paramslist]
+            new_params = list(paramslist)
+            for pc in part_cols:
+                if pc.lower() not in all_names and not should_hide_colname(pc):
+                    new_params.append(Param(pc, operators=["="], colname=pc))
+            return new_params
+    return paramslist
